feature-065: 카테고리 페이지 메뉴 부분 컴포넌트로 분리
